backend/pyproject.toml
[tool.poetry]
name = "leasing-agent"
version = "0.1.0"
description = ""
authors = ["ilyasubkhankulov <ilya.subkhankulov@gmail.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.12"}
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
pydantic-settings = {extras = ["dotenv"], version = "^2.9.1"}
nanoid = "^2.0.0"
asyncpg = "^0.30.0"
sqlmodel = "^0.0.24"
greenlet = "^3.2.2"
tenacity = "^9.1.2"
requests = "^2.32.4"
flake8-pyproject = "^1.2.3"
alembic = "^1.16.2"
openai = "^1.91.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
ignore = ["E203", "E501", "W503", "F401", "E731"]
exclude = [".git", "__pycache__", ".venv", "migrations"]